api/PromptFlow/QuestionAnswering/flows/standard/requirements.txt
promptflow_vectordb[azure]